Rename arches types file
