+++ arches/app/src/controlledlists/components/editor/AddValue.vue
@@ -14,7 +14,7 @@ import {
 } from "@/controlledlists/constants.ts";
 
 import type { Ref } from "vue";
-import type { Language } from "@/arches/types";
+import type { Language } from "@/types";
 import type {
     ControlledListItem,
     Value,
