+++ arches/app/src/controlledlists/components/misc/ListHeader.vue
@@ -10,7 +10,7 @@ import {
 import { bestLabel, dataIsList } from "@/controlledlists/utils.ts";
 
 import type { Ref } from "vue";
-import type { Language } from "@/arches/types";
+import type { Language } from "@/types";
 import type {
     ControlledList,
     DisplayedRowRefAndSetter,
