+++ arches/app/src/controlledlists/components/tree/ActionBanner.vue
@@ -10,7 +10,7 @@ import { bestLabel } from "@/controlledlists/utils.ts";
 import type { Ref } from "vue";
 import type { TreeSelectionKeys } from "primevue/tree/Tree";
 import type { TreeNode } from "primevue/treenode";
-import type { Language } from "@/arches/types";
+import type { Language } from "@/types";
 
 const isMultiSelecting = defineModel<boolean>("isMultiSelecting", {
     required: true,
