+++ arches/app/src/controlledlists/components/tree/AddDeleteControls.vue
@@ -22,7 +22,7 @@ import { dataIsItem, listAsNode } from "@/controlledlists/utils.ts";
 import type { Ref } from "vue";
 import type { TreeSelectionKeys } from "primevue/tree/Tree";
 import type { TreeNode } from "primevue/treenode";
-import type { Language } from "@/arches/types";
+import type { Language } from "@/types";
 import type {
     ControlledList,
     ControlledListItem,
