+++ arches/app/src/controlledlists/constants.ts
@@ -1,9 +1,9 @@
 import type { InjectionKey } from "vue";
+import type { Language } from "@/types";
 import type {
     ControlledList,
     ControlledListItem,
 } from "@/controlledlists/types";
-import type { Language } from "@/arches/types";
 
 export const displayedRowKey = Symbol() as InjectionKey<ControlledList | null>;
 export const itemKey = Symbol() as InjectionKey<ControlledListItem | null>;
